Add EventWaiter
The EventWaiter is a Discord/JDA event listener with an attached
ExecutorService that can be used to "wait" for specific events to be
fired, and return a Future back to the caller.

This is useful, for example, to implement things like timed response
wizards which require user input, reactions, etc before proceeding to
the next step.

There is a single event waiter per Alyx instance.