Move Alyx Plugins back into Alyx for the time being
I currently can't think of a way to "install" default
Alyx plugins into Alyx without making it so that other
third party plugins also have the ability to be automatically loaded,
which wouldn't be good. So for now, they'll live in Alyx and we'll
hardcode the ones we want to load initially until we have a better
solution.